+++ source/formats/gp7/parser.cpp
@@ -754,7 +754,7 @@ parseRhythms(const pugi::xml_node &rhythms_node)
 void
 Gp7::Document::addBar(MasterBar &master_bar, Bar bar)
 {
-    const int bar_id = myBars.size();
+    const int bar_id = static_cast<int>(myBars.size());
     myBars[bar_id] = std::move(bar);
     master_bar.myBarIds.push_back(bar_id);
 }
@@ -762,7 +762,7 @@ Gp7::Document::addBar(MasterBar &master_bar, Bar bar)
 void
 Gp7::Document::addVoice(Bar &bar, Voice voice)
 {
-    const int voice_id = myVoices.size();
+    const int voice_id = static_cast<int>(myVoices.size());
     myVoices[voice_id] = std::move(voice);
     bar.myVoiceIds.push_back(voice_id);
 }
@@ -770,7 +770,7 @@ Gp7::Document::addVoice(Bar &bar, Voice voice)
 void
 Gp7::Document::addBeat(Voice &voice, Beat beat)
 {
-    const int beat_id = myBeats.size();
+    const int beat_id = static_cast<int>(myBeats.size());
     myBeats[beat_id] = std::move(beat);
     voice.myBeatIds.push_back(beat_id);
 }
@@ -778,7 +778,7 @@ Gp7::Document::addBeat(Voice &voice, Beat beat)
 void
 Gp7::Document::addNote(Beat &beat, Note note)
 {
-    const int note_id = myNotes.size();
+    const int note_id = static_cast<int>(myNotes.size());
     myNotes[note_id] = std::move(note);
     beat.myNoteIds.push_back(note_id);
 }
@@ -787,7 +787,7 @@ void
 Gp7::Document::addRhythm(Beat &beat, Rhythm rhythm)
 {
     // TODO - consolidate identical rhythms?
-    const int rhythm_id = myRhythms.size();
+    const int rhythm_id = static_cast<int>(myRhythms.size());
     myRhythms[rhythm_id] = std::move(rhythm);
     beat.myRhythmId = rhythm_id;
 }
